+++ src/OpenLoco/src/GameCommands/GameCommands.h
@@ -653,129 +653,6 @@ namespace OpenLoco::GameCommands
         }
     };
 
-    struct TreeRemovalArgs
-    {
-        static constexpr auto command = GameCommand::removeTree;
-
-        TreeRemovalArgs() = default;
-        explicit TreeRemovalArgs(const registers& regs)
-            : pos(regs.ax, regs.cx, regs.dl * World::kSmallZStep)
-            , type(regs.dh)
-            , elementType(regs.bh)
-        {
-        }
-
-        World::Pos3 pos;
-        uint8_t type;
-        uint8_t elementType;
-
-        explicit operator registers() const
-        {
-            registers regs;
-            regs.ax = pos.x;
-            regs.cx = pos.y;
-            regs.dl = pos.z / World::kSmallZStep;
-            regs.dh = type;
-            regs.bh = elementType;
-            return regs;
-        }
-    };
-
-    struct TreePlacementArgs
-    {
-        static constexpr auto command = GameCommand::createTree;
-
-        TreePlacementArgs() = default;
-        explicit TreePlacementArgs(const registers& regs)
-            : pos(regs.ax, regs.cx)
-            , rotation(regs.di & 0x3)
-            , type(regs.bh)
-            , quadrant(regs.dl)
-            , colour(static_cast<Colour>(regs.dh & 0x1F))
-            , buildImmediately(regs.di & 0x8000)
-            , requiresFullClearance(regs.di & 0x4000)
-        {
-        }
-
-        World::Pos2 pos;
-        uint8_t rotation;
-        uint8_t type;
-        uint8_t quadrant;
-        Colour colour;
-        bool buildImmediately = false;
-        bool requiresFullClearance = false;
-
-        explicit operator registers() const
-        {
-            registers regs;
-            regs.ax = pos.x;
-            regs.cx = pos.y;
-            regs.dl = quadrant;
-            regs.dh = enumValue(colour);
-            regs.di = rotation | (buildImmediately ? 0x8000 : 0) | (requiresFullClearance ? 0x4000 : 0);
-            regs.bh = type;
-            return regs;
-        }
-    };
-
-    struct RaiseLandArgs
-    {
-        static constexpr auto command = GameCommand::raiseLand;
-        RaiseLandArgs() = default;
-        explicit RaiseLandArgs(const registers& regs)
-            : centre(regs.ax, regs.cx)
-            , pointA(regs.dx, regs.bp)
-            , pointB(regs.edx >> 16, regs.ebp >> 16)
-            , corner(regs.di)
-        {
-        }
-
-        World::Pos2 centre;
-        World::Pos2 pointA;
-        World::Pos2 pointB;
-        uint16_t corner;
-
-        explicit operator registers() const
-        {
-            registers regs;
-            regs.ax = centre.x;
-            regs.cx = centre.y;
-            regs.edx = (pointB.x << 16) | pointA.x;
-            regs.ebp = (pointB.y << 16) | pointA.y;
-            regs.di = corner;
-            return regs;
-        }
-    };
-
-    struct LowerLandArgs
-    {
-        static constexpr auto command = GameCommand::lowerLand;
-        LowerLandArgs() = default;
-        explicit LowerLandArgs(const registers& regs)
-            : centre(regs.ax, regs.cx)
-            , pointA(regs.dx, regs.bp)
-            , pointB(regs.edx >> 16, regs.ebp >> 16)
-            , corner(regs.di)
-        {
-        }
-
-        World::Pos2 centre;
-        World::Pos2 pointA;
-        World::Pos2 pointB;
-        uint16_t corner;
-
-        explicit operator registers() const
-        {
-            registers regs;
-            regs.ax = centre.x;
-            regs.cx = centre.y;
-            regs.edx = (pointB.x << 16) | pointA.x;
-            regs.ebp = (pointB.y << 16) | pointA.y;
-            regs.di = corner;
-            return regs;
-        }
-    };
-
     struct LowerRaiseLandMountainArgs
     {
         static constexpr auto command = GameCommand::lowerRaiseLandMountain;
@@ -801,115 +678,6 @@ namespace OpenLoco::GameCommands
             regs.edx = (pointB.x << 16) | pointA.x;
             regs.ebp = (pointB.y << 16) | pointA.y;
             regs.di = di;
-            return regs;
-        }
-    };
-
-    struct RaiseWaterArgs
-    {
-        static constexpr auto command = GameCommand::raiseWater;
-        RaiseWaterArgs() = default;
-        explicit RaiseWaterArgs(const registers& regs)
-            : pointA(regs.ax, regs.cx)
-            , pointB(regs.di, regs.bp)
-        {
-        }
-
-        World::Pos2 pointA;
-        World::Pos2 pointB;
-
-        explicit operator registers() const
-        {
-            registers regs;
-            regs.ax = pointA.x;
-            regs.cx = pointA.y;
-            regs.di = pointB.x;
-            regs.bp = pointB.y;
-            return regs;
-        }
-    };
-
-    struct LowerWaterArgs
-    {
-        static constexpr auto command = GameCommand::lowerWater;
-        LowerWaterArgs() = default;
-        explicit LowerWaterArgs(const registers& regs)
-            : pointA(regs.ax, regs.cx)
-            , pointB(regs.di, regs.bp)
-        {
-        }
-
-        World::Pos2 pointA;
-        World::Pos2 pointB;
-
-        explicit operator registers() const
-        {
-            registers regs;
-            regs.ax = pointA.x;
-            regs.cx = pointA.y;
-            regs.di = pointB.x;
-            regs.bp = pointB.y;
-            return regs;
-        }
-    };
-
-    struct WallPlacementArgs
-    {
-        static constexpr auto command = GameCommand::createWall;
-
-        WallPlacementArgs() = default;
-        explicit WallPlacementArgs(const registers& regs)
-            : pos(regs.ax, regs.cx, regs.di)
-            , rotation(regs.dl)
-            , type(regs.bh)
-            , primaryColour(static_cast<Colour>(regs.dh))
-            , secondaryColour(static_cast<Colour>(regs.bp & 0x1F))
-            , tertiaryColour(static_cast<Colour>((regs.bp >> 8) & 0x1F))
-        {
-        }
-
-        World::Pos3 pos;
-        uint8_t rotation;
-        uint8_t type;
-        Colour primaryColour;
-        Colour secondaryColour;
-        Colour tertiaryColour; // Note: will not work; render engine does not support tertiary
-
-        explicit operator registers() const
-        {
-            registers regs;
-            regs.ax = pos.x;
-            regs.cx = pos.y;
-            regs.dl = rotation;
-            regs.dh = enumValue(primaryColour);
-            regs.di = pos.z;
-            regs.bp = enumValue(secondaryColour) | (enumValue(tertiaryColour) << 8);
-            regs.bh = type;
-            return regs;
-        }
-    };
-
-    struct WallRemovalArgs
-    {
-        static constexpr auto command = GameCommand::removeWall;
-
-        WallRemovalArgs() = default;
-        explicit WallRemovalArgs(const registers& regs)
-            : pos(regs.ax, regs.cx, regs.dh * World::kSmallZStep)
-            , rotation(regs.dl)
-        {
-        }
-
-        World::Pos3 pos;
-        uint8_t rotation;
-
-        explicit operator registers() const
-        {
-            registers regs;
-            regs.ax = pos.x;
-            regs.cx = pos.y;
-            regs.dh = pos.z / World::kSmallZStep;
-            regs.dl = rotation;
             return regs;
         }
     };
@@ -1522,32 +1290,6 @@ namespace OpenLoco::GameCommands
         regs.bh = enumValue(company);
         return doCommand(GameCommand::changeCompanyFace, regs) != FAILURE;
     }
-
-    struct ClearLandArgs
-    {
-        static constexpr auto command = GameCommand::clearLand;
-        ClearLandArgs() = default;
-        explicit ClearLandArgs(const registers& regs)
-            : centre(regs.ax, regs.cx)
-            , pointA(regs.edx & 0xFFFF, regs.ebp & 0xFFFF)
-            , pointB(regs.edx >> 16, regs.ebp >> 16)
-        {
-        }
-
-        World::Pos2 centre;
-        World::Pos2 pointA;
-        World::Pos2 pointB;
-
-        explicit operator registers() const
-        {
-            registers regs;
-            regs.ax = centre.x;
-            regs.cx = centre.y;
-            regs.edx = (pointB.x << 16) | pointA.x;
-            regs.ebp = (pointB.y << 16) | pointA.y;
-            return regs;
-        }
-    };
 
     // Load multiplayer map
     inline void do_67(const char* filename)
