+++ src/OpenLoco/src/GameCommands/Industries/RenameIndustry.cpp
@@ -1,3 +1,4 @@
+#include "RenameIndustry.h"
 #include "Economy/Expenditures.h"
 #include "GameCommands/GameCommands.h"
 #include "Graphics/Gfx.h"
