+++ src/OpenLoco/src/GameCommands/Terraform/ClearLand.cpp
@@ -1,3 +1,4 @@
+#include "ClearLand.h"
 #include "Economy/Economy.h"
 #include "Economy/Expenditures.h"
 #include "GameCommands/GameCommands.h"
