+++ src/OpenLoco/src/GameCommands/Terraform/CreateTree.cpp
@@ -1,3 +1,4 @@
+#include "CreateTree.h"
 #include "Economy/Economy.h"
 #include "Economy/Expenditures.h"
 #include "GameCommands/GameCommands.h"
