+++ src/OpenLoco/src/GameCommands/Terraform/CreateWall.cpp
@@ -1,3 +1,4 @@
+#include "CreateWall.h"
 #include "Economy/Economy.h"
 #include "Economy/Expenditures.h"
 #include "GameCommands/GameCommands.h"
