+++ src/OpenLoco/src/GameCommands/Terraform/LowerLand.cpp
@@ -1,3 +1,4 @@
+#include "LowerLand.h"
 #include "Audio/Audio.h"
 #include "Economy/Expenditures.h"
 #include "GameCommands/GameCommands.h"
