+++ src/OpenLoco/src/GameCommands/Terraform/LowerWater.cpp
@@ -1,3 +1,4 @@
+#include "LowerWater.h"
 #include "Audio/Audio.h"
 #include "Economy/Expenditures.h"
 #include "GameCommands/GameCommands.h"
