+++ src/OpenLoco/src/GameCommands/Terraform/RaiseLand.cpp
@@ -1,3 +1,4 @@
+#include "RaiseLand.h"
 #include "Audio/Audio.h"
 #include "Economy/Expenditures.h"
 #include "GameCommands/GameCommands.h"
