+++ src/OpenLoco/src/GameCommands/Terraform/RaiseWater.cpp
@@ -1,3 +1,4 @@
+#include "RaiseWater.h"
 #include "Audio/Audio.h"
 #include "Economy/Expenditures.h"
 #include "GameCommands/GameCommands.h"
