+++ src/OpenLoco/src/Windows/TerraForm.cpp
@@ -3,6 +3,15 @@
 #include "Economy/Economy.h"
 #include "GameCommands/GameCommands.h"
 #include "GameCommands/Terraform/ChangeLandMaterial.h"
+#include "GameCommands/Terraform/ClearLand.h"
+#include "GameCommands/Terraform/CreateTree.h"
+#include "GameCommands/Terraform/CreateWall.h"
+#include "GameCommands/Terraform/LowerLand.h"
+#include "GameCommands/Terraform/LowerWater.h"
+#include "GameCommands/Terraform/RaiseLand.h"
+#include "GameCommands/Terraform/RaiseWater.h"
+#include "GameCommands/Terraform/RemoveTree.h"
+#include "GameCommands/Terraform/RemoveWall.h"
 #include "Graphics/Colour.h"
 #include "Graphics/ImageIds.h"
 #include "Input.h"
