+++ src/OpenLoco/src/World/Industry.cpp
@@ -2,6 +2,7 @@
 #include "Date.h"
 #include "GameCommands/GameCommands.h"
 #include "GameCommands/Industries/RemoveIndustry.h"
+#include "GameCommands/Terraform/CreateWall.h"
 #include "IndustryManager.h"
 #include "Localisation/Formatting.h"
 #include "Localisation/StringIds.h"
